refactoring) RacingGame - controller, Car - Model로 역할 분담
