+++ src/main/java/mission1/racing/RacingGameMain.java
@@ -4,6 +4,6 @@ public class RacingGameMain {
     private static RacingGame racingGame = new RacingGame();
 
     public static void main(String[] args) {
-        racingGame.announceWinner();
+        racingGame.announceWinners();
     }
 }
