+++ src/main/java/mission1/racing/model/Car.java
@@ -1,8 +1,13 @@
 package mission1.racing.model;
 
+import mission1.racing.RacingRandom;
+import mission1.racing.view.RacingGameOutputView;
+
 public class Car {
     private String name;
     private int carPositions;
+
+    private RacingRandom random = new RacingRandom();
 
     public Car() {}
     public Car(String name) {
@@ -10,7 +15,15 @@ public class Car {
     }
 
     public void go() {
-        carPositions++;
+        int LIMIT = 4;
+
+        if (random.random() >= LIMIT) {
+            carPositions++;
+        }
+    }
+
+    public boolean isWinner(int maxPosition) {
+        return carPositions == maxPosition;
     }
 
     public int getCarPositions() {
